Revert "web-sys: Added full Streams WebIDL (#3046)"
This reverts commit e8a499c36b41e197c7a1bd95050c73a6148e3f07.
