--- crates/web-sys/src/features/gen_ReadableStreamType.rs
@@ -1,10 +1,0 @@
-#![allow(unused_imports)]
-use wasm_bindgen::prelude::*;
-#[wasm_bindgen]
-#[doc = "The `ReadableStreamType` enum."]
-#[doc = ""]
-#[doc = "*This API requires the following crate features to be activated: `ReadableStreamType`*"]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ReadableStreamType {
-    Bytes = "bytes",
-}
