+++ crates/web-sys/src/features/gen_TransformStream.rs
@@ -1,6 +1,7 @@
 #![allow(unused_imports)]
 use super::*;
 use wasm_bindgen::prelude::*;
+#[cfg(web_sys_unstable_apis)]
 #[wasm_bindgen]
 extern "C" {
     # [wasm_bindgen (extends = :: js_sys :: Object , js_name = TransformStream , typescript_type = "TransformStream")]
@@ -10,7 +11,11 @@ extern "C" {
     #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/TransformStream)"]
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `TransformStream`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://rustwasm.github.io/docs/wasm-bindgen/web-sys/unstable-apis.html)*"]
     pub type TransformStream;
+    #[cfg(web_sys_unstable_apis)]
     #[cfg(feature = "ReadableStream")]
     # [wasm_bindgen (structural , method , getter , js_class = "TransformStream" , js_name = readable)]
     #[doc = "Getter for the `readable` field of this object."]
@@ -18,7 +23,11 @@ extern "C" {
     #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/TransformStream/readable)"]
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `ReadableStream`, `TransformStream`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://rustwasm.github.io/docs/wasm-bindgen/web-sys/unstable-apis.html)*"]
     pub fn readable(this: &TransformStream) -> ReadableStream;
+    #[cfg(web_sys_unstable_apis)]
     #[cfg(feature = "WritableStream")]
     # [wasm_bindgen (structural , method , getter , js_class = "TransformStream" , js_name = writable)]
     #[doc = "Getter for the `writable` field of this object."]
@@ -26,6 +35,9 @@ extern "C" {
     #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/TransformStream/writable)"]
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `TransformStream`, `WritableStream`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://rustwasm.github.io/docs/wasm-bindgen/web-sys/unstable-apis.html)*"]
     pub fn writable(this: &TransformStream) -> WritableStream;
     #[cfg(web_sys_unstable_apis)]
     #[wasm_bindgen(catch, constructor, js_class = "TransformStream")]
