--- crates/web-sys/src/features/gen_UnderlyingSink.rs
@@ -1,92 +1,0 @@
-#![allow(unused_imports)]
-use super::*;
-use wasm_bindgen::prelude::*;
-#[wasm_bindgen]
-extern "C" {
-    # [wasm_bindgen (extends = :: js_sys :: Object , js_name = UnderlyingSink)]
-    #[derive(Debug, Clone, PartialEq, Eq)]
-    #[doc = "The `UnderlyingSink` dictionary."]
-    #[doc = ""]
-    #[doc = "*This API requires the following crate features to be activated: `UnderlyingSink`*"]
-    pub type UnderlyingSink;
-}
-impl UnderlyingSink {
-    #[doc = "Construct a new `UnderlyingSink`."]
-    #[doc = ""]
-    #[doc = "*This API requires the following crate features to be activated: `UnderlyingSink`*"]
-    pub fn new() -> Self {
-        #[allow(unused_mut)]
-        let mut ret: Self = ::wasm_bindgen::JsCast::unchecked_into(::js_sys::Object::new());
-        ret
-    }
-    #[doc = "Change the `abort` field of this object."]
-    #[doc = ""]
-    #[doc = "*This API requires the following crate features to be activated: `UnderlyingSink`*"]
-    pub fn abort(&mut self, val: &::js_sys::Function) -> &mut Self {
-        use wasm_bindgen::JsValue;
-        let r = ::js_sys::Reflect::set(self.as_ref(), &JsValue::from("abort"), &JsValue::from(val));
-        debug_assert!(
-            r.is_ok(),
-            "setting properties should never fail on our dictionary objects"
-        );
-        let _ = r;
-        self
-    }
-    #[doc = "Change the `close` field of this object."]
-    #[doc = ""]
-    #[doc = "*This API requires the following crate features to be activated: `UnderlyingSink`*"]
-    pub fn close(&mut self, val: &::js_sys::Function) -> &mut Self {
-        use wasm_bindgen::JsValue;
-        let r = ::js_sys::Reflect::set(self.as_ref(), &JsValue::from("close"), &JsValue::from(val));
-        debug_assert!(
-            r.is_ok(),
-            "setting properties should never fail on our dictionary objects"
-        );
-        let _ = r;
-        self
-    }
-    #[doc = "Change the `start` field of this object."]
-    #[doc = ""]
-    #[doc = "*This API requires the following crate features to be activated: `UnderlyingSink`*"]
-    pub fn start(&mut self, val: &::js_sys::Function) -> &mut Self {
-        use wasm_bindgen::JsValue;
-        let r = ::js_sys::Reflect::set(self.as_ref(), &JsValue::from("start"), &JsValue::from(val));
-        debug_assert!(
-            r.is_ok(),
-            "setting properties should never fail on our dictionary objects"
-        );
-        let _ = r;
-        self
-    }
-    #[doc = "Change the `type` field of this object."]
-    #[doc = ""]
-    #[doc = "*This API requires the following crate features to be activated: `UnderlyingSink`*"]
-    pub fn type_(&mut self, val: &::wasm_bindgen::JsValue) -> &mut Self {
-        use wasm_bindgen::JsValue;
-        let r = ::js_sys::Reflect::set(self.as_ref(), &JsValue::from("type"), &JsValue::from(val));
-        debug_assert!(
-            r.is_ok(),
-            "setting properties should never fail on our dictionary objects"
-        );
-        let _ = r;
-        self
-    }
-    #[doc = "Change the `write` field of this object."]
-    #[doc = ""]
-    #[doc = "*This API requires the following crate features to be activated: `UnderlyingSink`*"]
-    pub fn write(&mut self, val: &::js_sys::Function) -> &mut Self {
-        use wasm_bindgen::JsValue;
-        let r = ::js_sys::Reflect::set(self.as_ref(), &JsValue::from("write"), &JsValue::from(val));
-        debug_assert!(
-            r.is_ok(),
-            "setting properties should never fail on our dictionary objects"
-        );
-        let _ = r;
-        self
-    }
-}
-impl Default for UnderlyingSink {
-    fn default() -> Self {
-        Self::new()
-    }
-}
