+++ crates/web-sys/src/features/mod.rs
@@ -718,12 +718,6 @@ mod gen_BrowserFindDirection;
 #[cfg(feature = "BrowserFindDirection")]
 pub use gen_BrowserFindDirection::*;
 
-#[cfg(feature = "ByteLengthQueuingStrategy")]
-#[allow(non_snake_case)]
-mod gen_ByteLengthQueuingStrategy;
-#[cfg(feature = "ByteLengthQueuingStrategy")]
-pub use gen_ByteLengthQueuingStrategy::*;
-
 #[cfg(feature = "Cache")]
 #[allow(non_snake_case)]
 mod gen_Cache;
@@ -1161,12 +1155,6 @@ pub use gen_ConvolverOptions::*;
 mod gen_Coordinates;
 #[cfg(feature = "Coordinates")]
 pub use gen_Coordinates::*;
-
-#[cfg(feature = "CountQueuingStrategy")]
-#[allow(non_snake_case)]
-mod gen_CountQueuingStrategy;
-#[cfg(feature = "CountQueuingStrategy")]
-pub use gen_CountQueuingStrategy::*;
 
 #[cfg(feature = "Credential")]
 #[allow(non_snake_case)]
@@ -5668,12 +5656,6 @@ mod gen_QueuingStrategy;
 #[cfg(feature = "QueuingStrategy")]
 pub use gen_QueuingStrategy::*;
 
-#[cfg(feature = "QueuingStrategyInit")]
-#[allow(non_snake_case)]
-mod gen_QueuingStrategyInit;
-#[cfg(feature = "QueuingStrategyInit")]
-pub use gen_QueuingStrategyInit::*;
-
 #[cfg(feature = "RadioNodeList")]
 #[allow(non_snake_case)]
 mod gen_RadioNodeList;
@@ -5698,17 +5680,17 @@ mod gen_RcwnStatus;
 #[cfg(feature = "RcwnStatus")]
 pub use gen_RcwnStatus::*;
 
-#[cfg(feature = "ReadableByteStreamController")]
-#[allow(non_snake_case)]
-mod gen_ReadableByteStreamController;
-#[cfg(feature = "ReadableByteStreamController")]
-pub use gen_ReadableByteStreamController::*;
-
 #[cfg(feature = "ReadableStream")]
 #[allow(non_snake_case)]
 mod gen_ReadableStream;
 #[cfg(feature = "ReadableStream")]
 pub use gen_ReadableStream::*;
+
+#[cfg(feature = "ReadableStreamByobReadResult")]
+#[allow(non_snake_case)]
+mod gen_ReadableStreamByobReadResult;
+#[cfg(feature = "ReadableStreamByobReadResult")]
+pub use gen_ReadableStreamByobReadResult::*;
 
 #[cfg(feature = "ReadableStreamByobReader")]
 #[allow(non_snake_case)]
@@ -5716,17 +5698,11 @@ mod gen_ReadableStreamByobReader;
 #[cfg(feature = "ReadableStreamByobReader")]
 pub use gen_ReadableStreamByobReader::*;
 
-#[cfg(feature = "ReadableStreamByobRequest")]
+#[cfg(feature = "ReadableStreamDefaultReadResult")]
 #[allow(non_snake_case)]
-mod gen_ReadableStreamByobRequest;
-#[cfg(feature = "ReadableStreamByobRequest")]
-pub use gen_ReadableStreamByobRequest::*;
-
-#[cfg(feature = "ReadableStreamDefaultController")]
-#[allow(non_snake_case)]
-mod gen_ReadableStreamDefaultController;
-#[cfg(feature = "ReadableStreamDefaultController")]
-pub use gen_ReadableStreamDefaultController::*;
+mod gen_ReadableStreamDefaultReadResult;
+#[cfg(feature = "ReadableStreamDefaultReadResult")]
+pub use gen_ReadableStreamDefaultReadResult::*;
 
 #[cfg(feature = "ReadableStreamDefaultReader")]
 #[allow(non_snake_case)]
@@ -5746,23 +5722,11 @@ mod gen_ReadableStreamIteratorOptions;
 #[cfg(feature = "ReadableStreamIteratorOptions")]
 pub use gen_ReadableStreamIteratorOptions::*;
 
-#[cfg(feature = "ReadableStreamReadResult")]
-#[allow(non_snake_case)]
-mod gen_ReadableStreamReadResult;
-#[cfg(feature = "ReadableStreamReadResult")]
-pub use gen_ReadableStreamReadResult::*;
-
 #[cfg(feature = "ReadableStreamReaderMode")]
 #[allow(non_snake_case)]
 mod gen_ReadableStreamReaderMode;
 #[cfg(feature = "ReadableStreamReaderMode")]
 pub use gen_ReadableStreamReaderMode::*;
-
-#[cfg(feature = "ReadableStreamType")]
-#[allow(non_snake_case)]
-mod gen_ReadableStreamType;
-#[cfg(feature = "ReadableStreamType")]
-pub use gen_ReadableStreamType::*;
 
 #[cfg(feature = "ReadableWritablePair")]
 #[allow(non_snake_case)]
@@ -7768,18 +7732,6 @@ mod gen_TransformStream;
 #[cfg(feature = "TransformStream")]
 pub use gen_TransformStream::*;
 
-#[cfg(feature = "TransformStreamDefaultController")]
-#[allow(non_snake_case)]
-mod gen_TransformStreamDefaultController;
-#[cfg(feature = "TransformStreamDefaultController")]
-pub use gen_TransformStreamDefaultController::*;
-
-#[cfg(feature = "Transformer")]
-#[allow(non_snake_case)]
-mod gen_Transformer;
-#[cfg(feature = "Transformer")]
-pub use gen_Transformer::*;
-
 #[cfg(feature = "TransitionEvent")]
 #[allow(non_snake_case)]
 mod gen_TransitionEvent;
@@ -7857,18 +7809,6 @@ pub use gen_UiEvent::*;
 mod gen_UiEventInit;
 #[cfg(feature = "UiEventInit")]
 pub use gen_UiEventInit::*;
-
-#[cfg(feature = "UnderlyingSink")]
-#[allow(non_snake_case)]
-mod gen_UnderlyingSink;
-#[cfg(feature = "UnderlyingSink")]
-pub use gen_UnderlyingSink::*;
-
-#[cfg(feature = "UnderlyingSource")]
-#[allow(non_snake_case)]
-mod gen_UnderlyingSource;
-#[cfg(feature = "UnderlyingSource")]
-pub use gen_UnderlyingSource::*;
 
 #[cfg(feature = "Url")]
 #[allow(non_snake_case)]
@@ -8685,12 +8625,6 @@ pub use gen_WorkletOptions::*;
 mod gen_WritableStream;
 #[cfg(feature = "WritableStream")]
 pub use gen_WritableStream::*;
-
-#[cfg(feature = "WritableStreamDefaultController")]
-#[allow(non_snake_case)]
-mod gen_WritableStreamDefaultController;
-#[cfg(feature = "WritableStreamDefaultController")]
-pub use gen_WritableStreamDefaultController::*;
 
 #[cfg(feature = "WritableStreamDefaultWriter")]
 #[allow(non_snake_case)]
